requirements.txt
anthropic
openai
python-dotenv
python-docx
